+++ rust/dbn/src/enums.rs
@@ -853,6 +853,7 @@ pub enum SecurityUpdateAction {
 #[repr(u16)]
 pub enum StatType {
     /// The price of the first trade of an instrument. `price` will be set.
+    /// `quantity` will be set when provided by the venue.
     OpeningPrice = 1,
     /// The probable price of the first trade of an instrument published during pre-
     /// open. Both `price` and `quantity` will be set.
@@ -884,6 +885,7 @@ pub enum StatType {
     /// set.
     FixingPrice = 10,
     /// The last trade price during a trading session. `price` will be set.
+    /// `quantity` will be set when provided by the venue.
     ClosePrice = 11,
     /// The change in price from the close price of the previous trading session to the
     /// most recent trading session. `price` will be set.
@@ -898,6 +900,10 @@ pub enum StatType {
     /// The option delta associated with the settlement price. `price` will be set with
     /// the standard precision.
     Delta = 15,
+    /// The auction uncrossing price. This is used for auctions that are neither the
+    /// official opening auction nor the official closing auction. `price` will be set.
+    /// `quantity` will be set when provided by the venue.
+    UncrossingPrice = 16,
 }
 
 /// The type of [`StatMsg`](crate::record::StatMsg) update.